requirements.txt
python-telegram-bot==12.0.0
python-dotenv==0.10.3